requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
seaborn
scikit-learn
pillow
opencv-python
tqdm
